DSE-28597: Set-up public git repo for cmlutils.
This is the initial commit for cmlutils(a command line utility tool for auxillary operations on CDSW/CML workspaces)
The tool initially aims to help simplify individual/batch migration of CML projects from one workspace to another. However, the tool might be extended to support many functionalities in the future.

Following are the important files that we are shipping with this initial commit:
 - README.md (This contains all the documentation to run the project-migration functionality till we find a suitable place for docs to live).
 - LICENSE
 - cmlutils/ (This directory contains the python source code for the tool)
 - examples/ (Sample code for batch migrating multiple projects at once. These build upon existing code which migrates one project at a time.)

Hello World ! :)