DSE-29039 Dynamic Population of engine->runtime mapping (#8)
* DSE-29039 Dynamic Population of engine->runtime mapping

This change helps in populating the Legacy engine to Target Runtime mapping on demand. Meaning, the runtime versions existing in the  Target Workspace will be automatically picked and used by the mapping.

* DSE-29039 Dynamic Population of engine->runtime mapping

This change helps in populating the Legacy engine to Target Runtime mapping on demand. Meaning, the runtime versions existing in the  Target Workspace will be automatically picked and used by the mapping.

* adding logger in the flow

* adding logger in the flow

* Writing engine-runtime map to json file, changing default runtime access