DSE-29903 - Jobs with same name are not getting migrated (#15)
For the jobs, there is no restriction on the workspace side. Two jobs can have same name and same parameters as well. During migration, we have added the check of not migrating jobs with same name to avoid duplicate jobs in the target workspace specially in cases of retry operation.
If we remove this check we might end up creating duplicate jobs, which is not desirable.
So we have added one more parameter "scriptname" along with job name during duplicate validation. This will allow certain jobs with same name to be migrated, but beyond this it is not feasible.

Co-authored-by: nagrawal <nagrawal@cloudera.com>